+++ app/[locale]/about/page.module.scss
