+++ app/[locale]/layout.tsx
@@ -1,0 +1,41 @@
+import type { Metadata } from "next";
+import { Inter } from "next/font/google";
+import Header from "@/components/header/header";
+import Footer from "@/components/footer/footer";
+import ContentWrapper from "@/components/contentWrapper/contentWrapper";
+import { NextIntlClientProvider } from "next-intl";
+import { notFound } from "next/navigation";
+
+const inter = Inter({ subsets: ["latin"] });
+
+export const metadata: Metadata = {
+  title: "Troinez",
+  description: "Generated by create next app",
+};
+
+export default async function RootLayout({
+  children,
+  params: { locale },
+}: {
+  children: React.ReactNode,
+  params: any,
+}) {
+  let messages;
+  try {
+    messages = (await import(`../../locales/${locale}.json`)).default;
+  } catch (error) {
+    notFound();
+  }
+
+  return (
+    <html lang={locale}>
+      <body className={inter.className} id="bodytest">
+        <NextIntlClientProvider locale={locale} messages={messages}>
+          <Header />
+          <ContentWrapper content={children} />
+          <Footer />
+        </NextIntlClientProvider>
+      </body>
+    </html>
+  );
+}
