--- app/about/page.tsx
@@ -1,9 +1,0 @@
-import styles from "./page.module.scss";
-
-export default function About() {
-  return (
-    <div>
-      <span className={styles.test}>어바웃</span>
-    </div>
-  );
-}
