--- app/page.tsx
@@ -1,7 +1,0 @@
-export default function Home() {
-  return (
-    <>
-      <a>홈</a>
-    </>
-  );
-}
